Estilizei a pagina LivrosDisponiveis, para listar os livros um de baixo do outro. Criei metodos na dao para retornar todos os livros e um para retornar somente a capa. Criei um serviço que devolve como resposta a capa do livro.

+++ MangaStore/DAO/IDataAccessObject.cs
@@ -13,7 +13,7 @@ namespace MangaStore.DAO
         /// </summary>
         /// <param name="cod"></param>
         /// <returns></returns>
-        List<object> Select(int cod, DataBaseHelper.SelectType selectType, DataBaseHelper.OrderBy orderBy);
+        List<object> Select(long cod, DataBaseHelper.SelectType selectType, DataBaseHelper.OrderBy orderBy);
 
         /// <summary>
         /// Realiza a inserção de dados
